requirements.txt
streamlit==1.7.0
h11==0.12.0
numpy==1.21.5
pandas==1.4.1
requests==2.26.0
plotly==5.5.0
Pillow==9.0.0